+++ examples/next-login/README.md
@@ -1,4 +1,4 @@
-# Ultraviolet UI Next Simple Example
+# Ultraviolet UI Next Login Example
 
 This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).
